[MOV] im_livechat: move LivechatButton outside of public_livechat.js
Task-2818675
